Phase A COMPLETE: Zero-Warning Build Achieved (12→0 warnings)
MILESTONE: Professional zero-warning build achieved

Warnings Fixed (12 → 0):

Group 1: Type Mismatch (1 warning) ✅
- clock_lattice.c/h: Changed float* to double* for consistency

Group 2: Format String Issues (4 warnings) ✅
- cllm_lattice_embeddings_spheres.c: Fixed %.2.0 → %.2f (lines 255, 394)

Group 3: SIMD Pointer Type Issues (5 warnings) ✅
- cllm_neighbor_ops.c: Disabled SIMD code temporarily (TODO: implement double-precision SIMD)
- Falls back to correct scalar implementation

Group 4: Struct Declaration (2 warnings) ✅
- cllm_threading.h: Removed unnecessary 'struct' keyword

Group 5: Unused Variable (1 warning) ✅
- cllm_neighbor_ops.c: Added __attribute__((unused)) for 'point'

Build Status:
- Errors: 0 ✅
- Warnings: 0 ✅ (down from 12)
- All libraries: Built successfully
- All tools: Built successfully

Impact:
- Professional code quality
- Bug prevention (type safety)
- CI/CD ready (can enforce zero-warning policy)
- Maintainability improved

Files Modified (5 files):
1. src/geometry/clock_lattice.c
2. include/clock_lattice.h
3. src/ai/cllm_lattice_embeddings_spheres.c
4. src/ai/cllm_neighbor_ops.c
5. include/cllm_threading.h

Documentation:
- Created ZERO_WARNINGS_ACHIEVED.md

Next: Phase B - Comprehensive Test Suite (167 tests)

+++ include/clock_lattice.h
@@ -80,7 +80,7 @@ bool is_sacred_position(BabylonianClockPosition pos);
  * @param coords Output: 3D coordinates [x, y, z]
  */
 void map_token_to_clock_lattice(uint32_t token_id, uint64_t prime, 
-                                int prime_index, float* coords);
+                                int prime_index, double* coords);
 
 /**
  * Get ring radius for visualization
